Reuse TestWithPiwikEnvironment in TestWithPiwikDatabase and fix some test issues.

+++ tests/PHPUnit/Framework/TestAspect/TestWithContainer.php
@@ -62,6 +62,8 @@ class TestWithContainer extends TestAspect
 
     public function destroyEnvironment()
     {
-        $this->environment->destroy();
+        if ($this->environment != null) {
+            $this->environment->destroy();
+        }
     }
 }
